requirements.txt
pywebview>=5.0
requests>=2.31
feedparser>=6.0
